Implement first branch (bne)
Not too sure I'm happy with the code for this yet, but we'll see if anything else has to be special-cased and see if patterns emerge as I move further. I'm concerned with trying too hard to find a good pattern before too many op's are implemented; who knows what surprises we'll run into down the line?