Impl mul ops
I'm actually a bit unsure how some of this should work; for example, do we set the sign/zero flags with the 64-bit result or the 32-bit low result (which I ended up choosing)? Is the handling of the 64-bit result correct for the 32-bit signed mul in this case? I guess we'll find out :)