Slow down gamepad strobe hack
Allows me to see stuff a bit clearer. That's a super random value btw; if it aligns too closely with the frame period the games won't catch it, so it has to be more or less random. The particular value was just some trial-and-error to find a decent 'blast through menus' speed for now :)