Move frame cycle tracking to VirtualBoy
I'm a bit torn on this, but I think it will clean stuff up and allow us to easily add a 'step frame' command, which will be rad. The reason I don't quite like it, however, is I'm not sure if 'run the emu one frame' will always be an appropriate interface, particularly when we get sounds up and running, but I think we should go with this for now and worry about that later.