Only swap framebuffers when drawing is enabled
This fixes the intro animation of one of the games. It's a bit counterintuitive to swap buffers _before_ drawing rather than after, but I tried doing so after instead and it didn't play very nicely. That might point to something more subtle, but I think this is a good fix for now at least.