Refactor draw_pixel -> draw_char_pixel
All of the drawing ops were doing more or less the same character reads/pixel writes. This moves some of the address resolution into the innermost loops, but I'm assuming for now it won't make much difference. Depending on inlining this could be alleviated (due to lifting), so it might be a good idea to add inline hints later, but I don't want to touch that until the code is more readible/simple and I've actually determined this is a perf problem in the first place.