Enable display on startup and only fire start of frame in while display is enabled
So this was a bit crap, since all of the docs claim the display hardware is disabled by default. It's the only way to make some games work, however, once the start of frame interrupts only trigger while the display is active (which is a requirement for other games to work :P).

I'm guessing the display actually _could_ be active on hw reset, as you still wouldn't get any picture without the sync signals from the left/right scanners active, so this is sensible enough I suppose.

I was trying endless things until finally I dug into mednafen, and they had this state enabled on startup. I'm still not 100% sure it's correct, but I don't have the hw to fully test, and it DOES work, so I'm pushing it. rrrg.