Early out breakpoint + watchpoint checks
Did a bit of profiling and found out the emu was spending an absurd amount of time doing hashset lookups for breakpoints/watchpoints. Dug into the rust std source and noticed these lookups don't early-out if the collection is empty, which in this scenario is the most common case by far, so this was an easy change.