Add feature for logging
This covers the feature itself and replacing all of the relevant print/println's with log/logln's. Doesn't cover accounting for new warnings as a result of the code that's removed when the feature isn't enabled (and it's not by default).

This is useful because the output is currently invaluable for debugging, but on many systems, can severely hurt performance. An end-user wouldn't have any use for the log output, so it shouldn't be there by default. While I would normally hold off on a change like this as I consider it polish, it's also affecting my perf profiling, making it more difficult to weed out important information, so doing this now makes some sense.