Refactor instruction cycle counting pattern
Although this moves the instruction cycle counts out of the instruction
module and spreads them out a bit over the nvc code, the overall pattern
is a bit simpler and it's a tiny bit faster. In the common case (most
instructions take a single cycle to execute) this doesn't change the
code at all.