Don't panic on unrecognized system reg index
For some reason the startup code of a bunch of homebrew rom's want to load a number into system register 14, which doesn't map to anything as far as I know. I double-checked that this disassembly was correct (by hand :P) and it appeared to be, so this is a bit odd.