Removed per-sample locking
Struggled a bit to find a good abstraction and had a good fight with the borrow-checker. In the end I decided to stop trying to abstract and just get it to work, even though I've exposed some locking to the top-level code. Not terribly much though, and some things still got simpler up there. Overall less code again at least :)