Remove DC offset from VSU output
This also means scaling it down, so it's not as loud, but the reason for this is that on systems that are too slow to keep up (which is common right now as the emu is still quite unoptimized) the audio buffer will often be empty. Since we're using a VecDeque for this, we just output 0 for the samples that aren't filled in (rather than whatever would be in the ring buffer otherwise). With a DC offset, this creates very disturbing clicks/pops, esp. since the DC offset was max volume.