Move duration+envelope clock counters to vsu
No need for them to be per-voice, as the real hw would likely have a single clock divider circuit that would feed into all of them.

I'm considering doing some iterators for the voices, as there are an increasing number of code blocks that just call some fn for each voice, but I have a feeling that would negatively affect performance.