Adjust small timer interval
This fixes the vocal sample play rate in galactic pinball, although
the timing conflicts a bit with the docs. See comment for more info.

I also considered the possibility that the timer algorithm itself is
off, but the timer appears to behave correctly with this algorithm with
the larger interval, and tweaking it seems to cause drift in the
recordings I took, so I don't think that's the case. It's especially
good evidence that the value, while not matching the docs, _does_ end up
being a sensible, round number.