Adjust small timer interval 333 -> 332 clocks
I keep going back and forth with this, but the logic here makes sense. In my recordings, I've tried to have this interval at 332 and 333 clock cycles. At 332, the sample in Galactic Pinball plays _slightly faster_ than the one on the hw (not noticably, but when comparing in a DAW); at 333, it plays _slightly slower_. It looks like the hardware is right in the middle of these two intervals. But, remember that I'm rounding the sample rate of the emulator up to the nearest whole-number of samples per second, which _should_ mean that the recording from the emu should actually be _very slightly faster_ than the recording from the hardware, which matches the 332 clocks period rather than the 333 clocks period. Something about it being an even number also feels more consistent with other counter values in the emu.