Disable timer interrupt if zero status is cleared
Before it would only disable if interrupt enable was cleared, but it seems Virtual Bowling only clears zero status and expects the interrupt to be ACK'd, so we'll support both. This appears to be consistent with the documentation which recommends clearing this flag as well as disabling the timer interrupt, and doesn't appear to have any negative side-effects.