Adjust small timer interval again
I hadn't tested Vertical Force in awhile, and noticed it its music was playing noticeably too fast. I double-checked on hw, and ofc, it certainly was. Turns out, putting the timer interval back at 20us like the docs say fixes it. Ugh.

I'm still putting my money re the sample playback speeds in other games being wrong on the cpu not being cycle-accurate and thus not enabling/disabling the timer as tightly as it could be. Need to dig into those routines to be sure, but hey, at least we have some conclusive evidence that this timer interval is what the docs claim.