adjust time source start time when coming out of the debugger
When in the debugger the time source will continue to progress, while emulation
time is mostly standing still.  When you exit the debugger the emulator
tries to make emulation time catch up with the time source by running a very
large number of instructions very fast.

To work around this we can adjust the time source start time to be the time
gap between the time source and the emulation time, which will make the 2
times even again.  Thus causing emulation to run at normal speeds