Add the ability to enable/disable logging of each subsystem in the core
log(ln)! macros in the core were update to accept additional formats that
support the first param being the Log enum.  The existing 'logging' feature
in the core was replaced with the follow features for the different subsystems.

log-cpu
log-gamepad
log-ic
log-vip
log-vsu
log-other (anything not passing the Log enum to the log(ln)! macros)
log-all

Since the core is a dependency of the cli it was necessary to add features
the cli's Cargo.toml to pass along enabling the log feature(s) to the core.
The following names were used for this in the cli.

log-core-cpu
log-core-gamepad
log-core-ic
...

The cli's 'logging' feature was renamed to 'log-cli' to give it the same
naming convention as the core.  The log(ln)! macros in the cli were also
updated to be the same style as the core, minus the Log enum bits.