Quick and dirty startup display stabilization
This simulates the display hw not being ready on startup until the mirror oscillation stabilizes. This is an attempt to fix Space Squash not starting, but doesn't seem to be very effective. I need to debug some more, but it would probably be a good idea to emulate this with a command line option for how long this period should last.

Also, this implementation doesn't correctly output dummy frames during this period, so it looks like the emu hangs. This should be fixed if this feature becomes more mature.