Add cpu tracing thats enabled via features cpu-trace
cpu tracing allows tracking the before and after states of the
registers/memory/etc that an instruction will use.  This before and
after data can then be displayed when using the step function in the
debugger.

(vb-rs 0xfff90fea) > cputrace
cpu trace enabled: true

Then when you run the step command it will look like this

(vb-rs 0x07000c00) > s
  0x07000c00  0fce0400    ld.w 4[r15], r16
    before PC: 0x07000c00 PSW: 0x00010000 r15: 0x0500085c r16: 0x00000000 disp: 4  mem.w: [addr: 0x05000860, value: 0x0500089c] cpu cycles: 4
     after PC: 0x07000c04 PSW: 0x00010000 r15: 0x0500085c r16: 0x0500089c disp: 4  mem.w: [addr: 0x05000860, value: 0x0500089c]

This required tweaking the debugger step function which used to
cpu.step(), then disassemble() to disassemble(), then cpu.step().  This
was needed to make the pc reg and the disassembled instruction match up.