Account for cpu cycles that occur when entering an exception
I wasnt able to find any docs that indicate how many cpu cycles it takes
to enter an exception.  I ended up going with 10, which is how many it
takes to leave an exception (reti).  However I suspect entering an
exception is really going to be a few more then that.

The added cpu cycles are applied to the first instruction executed in
the exception.