Implement FEPC/FEPSW for ldsr/stsr
Only loads/stores the values; duplexed exceptions are still not currently implemented. The masking here was found during simple hardware tests, and it makes sense that the lowest bit may not be present in the hardware, as instructions are always 2 or 4 bytes long, meaning the lowest bit of PC/EIPC/FEPC will always be zero. This is implemented by just masking the values on writes (as well as initial values).