Output in 0rgb1555
This is the default libretro format, and since retroarch seems to ignore our pixel format call sometimes, we should just output in this format for now. Eventually I'd like to support all possible formats and let the host give us the output buffer, but for now this makes things a bit more stable.