SQUASHME: More serialization progress
Decided to stop hiding internal buffers, and cleaning up associated code. Now we also avoid reallocating a bunch of buffers (particularly ROM and SRAM) when new data comes in, which means both that the pointers that the libretro impl gives to the frontend will never move, and we get to remove a lot of data hiding code that we didn't need anyways. I'd like basically everything in the core itself to be public so that any tooling that wants to inspect the internal state can without any fluff.