Use 12 bits only for background x and parallax
This is documented in the hardware, but since they're stored in DRAM, I just treated them as 16 bits before. Enforcing this fixes Space Invaders - the last game before we reach 100% compatibility :)

Something to note here is this _only_ covers the x and parallax window attributes; none of the other ones. Before enforcing size on any of the other ones I'd like to come up with some video hardware fuzzing tests.